relocate loginpage queries
